+++ cities_indicators/city.py
@@ -84,7 +84,7 @@ class City:
 
     def get_geom(self, admin_level):
         unit_geom = requests.get(f"{API_URI}/{self.id}/{admin_level}/geojson").json()
-        unit_boundaries = gpd.GeoDataFrame.from_features(unit_geom)
-        unit_boundaries["index"] = unit_boundaries["geo_id"].apply(lambda x: int(x.split("_")[2]) - 1)
-        unit_boundaries = unit_boundaries.sort_values(by="index")
+        unit_boundaries = gpd.GeoDataFrame.from_features(unit_geom).reset_index()
+        # unit_boundaries["index"] = unit_boundaries["geo_id"].apply(lambda x: int(x.split("_")[2]) - 1)
+        # unit_boundaries = unit_boundaries.sort_values(by="index")
         return unit_boundaries
