+++ tests/cities_indicators_test.py
@@ -1,3 +1,6 @@
+import coiled
+from distributed import Client
+
 from cities_indicators.core import get_city_indicators, get_indicators, Indicator
 from cities_indicators.city import get_city_admin, API_URI
 
@@ -23,9 +26,9 @@ def test_tree_cover_in_built_up_areas():
 
 
 def test_surface_reflectivity():
-    jakarta = get_city_admin("IDN-Jakarta")[1:]
+    # create a remote Dask cluster with Coiled
+    jakarta = get_city_admin("ARG-Buenos_Aires")[1:]
     indicators = get_city_indicators(cities=jakarta, indicators=[Indicator.SURFACE_REFLECTIVTY])[0]
-    baseline_indicators = get_baseline("HEA_3_percentBuiltwLowAlbedo")
 
     for actual, baseline in zip(indicators["HEA_3_percentBuiltwLowAlbedo"], ACTUAL_ALBEDO_VALUES):
         assert pytest.approx(actual, abs=0.015) == baseline
