+++ tests/cities_layers_test.py
@@ -1,7 +1,7 @@
 #from cities_indicators.core import get_indicators, Indicator
 from cities_indicators.city import SupportedCity, City
-from cities_indicators.layers.albedo import Albedo
-from cities_indicators.layers.land_surface_temperature import LandSurfaceTemperature
+from cities_indicators.layers.albedo_gee import AlbedoGEE
+from cities_indicators.layers.land_surface_temperature_gee import LandSurfaceTemperatureGEE
 from cities_indicators.indicators.built_land_with_high_lst import BuiltUpHighLandSurfaceTemperature
 from cities_indicators.indicators.built_land_with_high_lst_gee import BuiltUpHighLandSurfaceTemperatureGEE
 # city
